+++ packages/system/src/components/Global/Global.tsx
@@ -1,39 +1,33 @@
 import React from 'react';
 import { Global as EmotionGlobal } from '@emotion/react';
-import { useTheme } from '../../hooks';
 
-/**
- * CSS snippet and idea from:
- * https://css-tricks.com/revisiting-prefers-reduced-motion-the-reduced-motion-media-query/
- */
-const reduceMotionStyles = {
-  '@media screen and (prefers-reduced-motion: reduce), (update: slow)': {
-    '*': {
-      animationDuration: '0.001ms !important',
-      animationIterationCount: '1 !important',
-      transitionDuration: '0.001ms !important',
-    },
-  },
+import { useTheme } from '../../hooks/useTheme';
+import * as normalize from './normalize';
+
+export type GlobalProps = {
+  /**
+   * CSS Selector to change the element that the styles will be applied to.
+   */
+  selector?: string;
+
+  /**
+   * Normalize `html` and `body`? Defaults to `true`.
+   */
+  normalizeDocument?: boolean;
 };
 
-export const Global = () => {
-  const { css } = useTheme();
-  const styles = css({
-    html: {
-      height: '100%',
-      /**
-       * Prevent Mobile Safari from zooming stuff ...
-       * Source: https://css-tricks.com/your-css-reset-needs-text-size-adjust-probably/
-       */
-      textSizeAdjust: 'none',
-      variant: 'root.html',
-    },
-    body: {
-      height: '100%',
-      lineHeight: 1.5,
-      WebkitFontSmoothing: 'antialiased',
-      variant: 'root.body',
-    },
-  });
-  return <EmotionGlobal styles={{ reduceMotionStyles, ...styles }} />;
+export const Global = ({ normalizeDocument = true, selector }: GlobalProps) => {
+  const { css, theme } = useTheme();
+  const rootStyles = css(theme?.root || {});
+
+  const styles = [
+    normalizeDocument ? normalize.document : {},
+    // Prefix normalization and root styles with selector if provided.
+    selector
+      ? { [`:where(${selector})`]: normalize.element }
+      : normalize.element,
+    selector ? { [`:where(${selector})`]: rootStyles } : rootStyles,
+  ];
+
+  return <EmotionGlobal styles={styles} />;
 };