+++ packages/system/src/types/theme.ts
@@ -113,6 +113,13 @@ export interface ZeroSizeScale<T> extends SizeScale<T> {
  */
 export interface Theme {
   /**
+   * Styles that applied to the root of the app, where the root can be either the document
+   * for full page apps or an element if you're only using Marigold on a portion of the page.
+   * This is useful for additional normalizationor and to style non-Marigold elements.
+   */
+  root?: CSSObject;
+
+  /**
    * To configure the default breakpoints used in responsive array values,
    * add a breakpoints array to your theme.
    *