+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @marigold/icons
 
+## 0.7.5
+
+### Patch Changes
+
+- [#2164](https://github.com/marigold-ui/marigold/pull/2164) [`b43464fc`](https://github.com/marigold-ui/marigold/commit/b43464fce6f7e2662b27313c6f74190e8c0f540f) Thanks [@sarahgm](https://github.com/sarahgm)! - refa: update to React 18
+
+- Updated dependencies [[`b43464fc`](https://github.com/marigold-ui/marigold/commit/b43464fce6f7e2662b27313c6f74190e8c0f540f)]:
+  - @marigold/system@1.2.1
+
+## 0.7.4
+
+### Patch Changes
+
+- Updated dependencies [[`82c376a9`](https://github.com/marigold-ui/marigold/commit/82c376a9c34c69104456e6e356231132f53c3efe)]:
+  - @marigold/system@1.2.0
+
+## 0.7.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @marigold/system@1.1.1
+
 ## 0.7.2
 
 ### Patch Changes